api_gateway/requirements.txt
uvicorn==0.34.2
fastapi==0.115.12
python-dotenv==1.1.0
grpcio==1.71.0
grpcio-reflection==1.71.0
protobuf==5.29.4